keeping these components for future reference. Moved into the sidebar-components folder for storage.
